+++ Aula2/src/ex2/Pessoa.java
@@ -45,7 +45,7 @@ public class Pessoa {
         return altura;
     }
 
-    public void setAltura(float altura) {
+    public void setAltura(double altura) {
         this.altura = altura;
     }
